Troca int por float notas

+++ faculdade/ads-uninter/logica-e-algoritmos/pratica/exercicio3.py
@@ -20,10 +20,10 @@ for aluno in range (qtd_alunos):
     cls()
     nome_aluno = input('Digite o nome do aluno: ')
 
-    nota1 = int(input('Insira a nota 1: '))
-    nota2 = int(input('Insira a nota 2: '))
-    nota3 = int(input('Insira a nota 3: '))
-    nota4 = int(input('Insira a nota 4: '))
+    nota1 = float(input('Insira a nota 1: '))
+    nota2 = float(input('Insira a nota 2: '))
+    nota3 = float(input('Insira a nota 3: '))
+    nota4 = float(input('Insira a nota 4: '))
     media = (nota1 + nota2 + nota3 + nota4)/4
 
     alunos['nome'].append(nome_aluno)
